structural: Centralize lifecycle test defaults
